backend/requirements.txt
# Core
Django==5.2.12
asgiref==3.11.1
sqlparse==0.5.5
typing_extensions==4.15.0

# Database
psycopg2-binary==2.9.10

# Cache
django-redis==5.4.0

# Server
gunicorn==23.0.0

# Environment
python-dotenv==1.1.0

# Testing
pytest==8.3.5
pytest-django==4.10.0
pytest-cov==6.1.1
factory-boy==3.3.3

# API & SPA
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
Pillow==11.1.0

# Tasks
celery==5.4.0
